requirements.txt
astroid==2.4.2
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
pylint==2.6.0
PyYAML==5.3.1
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
wrapt==1.12.1
